refactor: remove open youtube from view model
